+++ builddocs/conf.py
@@ -53,7 +53,7 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 html_theme = 'pydata_sphinx_theme'
 html_theme_options = {
     "github_url": "https://github.com/kconnour/pyRT_DISORT",
-    "collapse_navigation": True,
+    #"collapse_navigation": True,
 }
 
 # Add any paths that contain custom static files (such as style sheets) here,